pass java error to task

+++ helper.rb
@@ -103,28 +103,29 @@ module OpenTox
       def isa2rdf
         # @note isa2rdf returns correct exit code but error in task
         # @todo delete dir if task catches error, pass error to block
-        begin
-          `cd #{File.dirname(__FILE__)}/java && java -jar -Xmx2048m isa2rdf-cli-0.0.7.jar -d #{tmp} -o #{File.join tmp,nt} -t #{$user_service[:uri]} `#&> #{File.join tmp,'log'}`
-        rescue
+        `cd #{File.dirname(__FILE__)}/java && java -jar -Xmx2048m isa2rdf-cli-0.0.7.jar -d #{tmp} -o #{File.join tmp,nt} -t #{$user_service[:uri]} 2> #{File.join tmp,'log'} &`
+        if !File.exists?(File.join tmp, nt)
+          out = IO.read(File.join tmp, 'log') 
           FileUtils.remove_entry dir
           delete_investigation_policy
-          bad_request_error "Could not parse isatab file in #{params[:file][:filename]}."
+          bad_request_error "Could not parse isatab file in '#{params[:file][:filename]}'. Message is:\n #{out}"
+        else
+          `sed -i 's;http://onto.toxbank.net/isa/tmp/;#{investigation_uri}/;g' #{File.join tmp,nt}`
+          investigation_id = `grep "#{investigation_uri}/I[0-9]" #{File.join tmp,nt}|cut -f1 -d ' '`.strip
+          `sed -i 's;#{investigation_id.split.last};<#{investigation_uri}>;g' #{File.join tmp,nt}`
+          # `echo '\n<#{investigation_uri}> <#{RDF::DC.modified}> "#{Time.new.strftime("%d %b %Y %H:%M:%S %Z")}" .' >> #{File.join tmp,nt}`
+          `echo "\n<#{investigation_uri}> <#{RDF.type}> <#{RDF::OT.Investigation}> ." >>  #{File.join tmp,nt}`
+          FileUtils.rm Dir[File.join(tmp,"*.zip")]
+          FileUtils.cp Dir[File.join(tmp,"*")], dir
+          `zip -j #{File.join(dir, "investigation_#{params[:id]}.zip")} #{dir}/*.txt`
+          OpenTox::Backend::FourStore.put investigation_uri, File.read(File.join(dir,nt)), "application/x-turtle"
+          FileUtils.remove_entry tmp
+          newfiles = `cd #{File.dirname(__FILE__)}/investigation; git ls-files --others --exclude-standard --directory #{params[:id]}`
+          `cd #{File.dirname(__FILE__)}/investigation && git add #{newfiles}`
+          request.env['REQUEST_METHOD'] == "POST" ? action = "created" : action = "modified"
+          `cd #{File.dirname(__FILE__)}/investigation && git commit -am "investigation #{params[:id]} #{action} by #{OpenTox::Authorization.get_user}"`
+          investigation_uri
         end
-        `sed -i 's;http://onto.toxbank.net/isa/tmp/;#{investigation_uri}/;g' #{File.join tmp,nt}`
-        investigation_id = `grep "#{investigation_uri}/I[0-9]" #{File.join tmp,nt}|cut -f1 -d ' '`.strip
-        `sed -i 's;#{investigation_id.split.last};<#{investigation_uri}>;g' #{File.join tmp,nt}`
-        # `echo '\n<#{investigation_uri}> <#{RDF::DC.modified}> "#{Time.new.strftime("%d %b %Y %H:%M:%S %Z")}" .' >> #{File.join tmp,nt}`
-        `echo "\n<#{investigation_uri}> <#{RDF.type}> <#{RDF::OT.Investigation}> ." >>  #{File.join tmp,nt}`
-        FileUtils.rm Dir[File.join(tmp,"*.zip")]
-        FileUtils.cp Dir[File.join(tmp,"*")], dir
-        `zip -j #{File.join(dir, "investigation_#{params[:id]}.zip")} #{dir}/*.txt`
-        OpenTox::Backend::FourStore.put investigation_uri, File.read(File.join(dir,nt)), "application/x-turtle"
-        FileUtils.remove_entry tmp
-        newfiles = `cd #{File.dirname(__FILE__)}/investigation; git ls-files --others --exclude-standard --directory #{params[:id]}`
-        `cd #{File.dirname(__FILE__)}/investigation && git add #{newfiles}`
-        request.env['REQUEST_METHOD'] == "POST" ? action = "created" : action = "modified"
-        `cd #{File.dirname(__FILE__)}/investigation && git commit -am "investigation #{params[:id]} #{action} by #{OpenTox::Authorization.get_user}"`
-        investigation_uri
       end
 
       # creates XML policy file for user or group
